🎨refactor: Popup class 파일 위치 수정
